// src/lib/api/generated/model/maintenancePageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { MaintenanceProjectionSchema } from './maintenanceProjectionSchema';

export interface MaintenancePageSchema {
  items: MaintenanceProjectionSchema[];
  page: number;
  page_size: number;
  total: number;
}