// src/lib/api/generated/model/systemErrorPageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { ApplicationErrorProjectionSchema } from './applicationErrorProjectionSchema';

export interface SystemErrorPageSchema {
  items: ApplicationErrorProjectionSchema[];
  page: number;
  page_size: number;
  total: number;
}